Add Support for more precise coordinates

+++ choo/apis/efa.py
@@ -654,7 +654,7 @@ class EFA(API):
 
         path = []
         for coords in data.findall('./itdPathCoordinates/itdCoordinateBaseElemList/itdCoordinateBaseElem'):
-            path.append(Coordinates(int(coords.find('y').text) / 1000000, int(coords.find('x').text) / 1000000))
+            path.append(Coordinates(float(coords.find('y').text) / 1000000, float(coords.find('x').text) / 1000000))
 
         motdata = self._parse_mot(data.find('./itdMeansOfTransport'))
